Tree-wide int->ptrdiff_t index conversion driven to par<=ob parity.
See the merged commit for the full regression-class breakdown.

+++ src/epblas-parallel/kind10/egemm_kernel.h
@@ -34,34 +34,34 @@ typedef long double egemm_T;
 #define EGEMM_NR 2
 
 /* Normalize a Fortran trans char to a code ('C' ≡ 'T' for real input). */
-int egemm_trans_code(const char *p, size_t len);
+ptrdiff_t egemm_trans_code(const char *p, size_t len);
 
-int egemm_round_up(int v, int m);
+ptrdiff_t egemm_round_up(ptrdiff_t v, ptrdiff_t m);
 
 /* Cache-block sizes (env-overridable EBLAS_MC/KC/NC) with OpenBLAS-style
  * adaptive MC when K fits one panel. */
-void egemm_choose_blocks(int K, int *MC, int *KC, int *NC);
+void egemm_choose_blocks(ptrdiff_t K, ptrdiff_t *MC, ptrdiff_t *KC, ptrdiff_t *NC);
 
 /* C := beta*C pre-pass over the full M×N tile (handles K==0 / alpha==0). */
-void egemm_beta_prepass(int M, int N, egemm_T beta, egemm_T *c, int ldc);
+void egemm_beta_prepass(ptrdiff_t M, ptrdiff_t N, egemm_T beta, egemm_T *c, ptrdiff_t ldc);
 
 /* Packers (panel-packed, OpenBLAS-style). */
-void egemm_pack_A(const egemm_T *restrict A, int lda,
-                  int ic, int pc, int ib, int pb, int ta,
+void egemm_pack_A(const egemm_T *restrict A, ptrdiff_t lda,
+                  ptrdiff_t ic, ptrdiff_t pc, ptrdiff_t ib, ptrdiff_t pb, ptrdiff_t ta,
                   egemm_T *restrict Ap);
-void egemm_pack_B(const egemm_T *restrict B, int ldb,
-                  int pc, int jc, int pb, int jb, int tb,
+void egemm_pack_B(const egemm_T *restrict B, ptrdiff_t ldb,
+                  ptrdiff_t pc, ptrdiff_t jc, ptrdiff_t pb, ptrdiff_t jb, ptrdiff_t tb,
                   egemm_T *restrict Bp);
 
 /* Drive one packed (ib,jb,pb) macro-tile via MR×NR sub-tiles. */
-void egemm_macro_kernel(int ib, int jb, int pb, egemm_T alpha,
+void egemm_macro_kernel(ptrdiff_t ib, ptrdiff_t jb, ptrdiff_t pb, egemm_T alpha,
                         const egemm_T *restrict Ap, const egemm_T *restrict Bp,
-                        egemm_T *restrict C, int ldc);
+                        egemm_T *restrict C, ptrdiff_t ldc);
 
 /* Fast path TA='T',TB='N': one C-column j2 (stride-1 dot, no packing). */
-void egemm_fast_col(int j2, int M, int K, egemm_T alpha,
-                    const egemm_T *a, int lda, const egemm_T *b, int ldb,
-                    egemm_T *c, int ldc);
+void egemm_fast_col(ptrdiff_t j2, ptrdiff_t M, ptrdiff_t K, egemm_T alpha,
+                    const egemm_T *a, ptrdiff_t lda, const egemm_T *b, ptrdiff_t ldb,
+                    egemm_T *c, ptrdiff_t ldc);
 
 /*
  * Gate for the TN (ta='T', tb='N') no-pack fast_col path. fast_col runs the
@@ -72,19 +72,19 @@ void egemm_fast_col(int j2, int M, int K, egemm_T alpha,
  * any non-trivial K. fast_col only wins where packing isn't amortized:
  * short K (the single chain hasn't yet lost to the blocked ILP) or a tiny
  * output. Everything else falls through to the blocked path. */
-static inline int egemm_tn_use_fast(int M, int N, int K) {
+static inline ptrdiff_t egemm_tn_use_fast(ptrdiff_t M, ptrdiff_t N, ptrdiff_t K) {
     return K <= 64 || (long)M * (long)N <= 64L * 64L;
 }
 
 /* Pure single-thread GEMM. Same signature as egemm_ — no OpenMP. */
 void egemm_serial(
     const char *transa, const char *transb,
-    const int *m_, const int *n_, const int *k_,
+    const ptrdiff_t *m_, const ptrdiff_t *n_, const ptrdiff_t *k_,
     const egemm_T *alpha_,
-    const egemm_T *a, const int *lda_,
-    const egemm_T *b, const int *ldb_,
+    const egemm_T *a, const ptrdiff_t *lda_,
+    const egemm_T *b, const ptrdiff_t *ldb_,
     const egemm_T *beta_,
-    egemm_T *c, const int *ldc_,
+    egemm_T *c, const ptrdiff_t *ldc_,
     size_t transa_len, size_t transb_len);
 
 #endif /* EPBLAS_PARALLEL_KIND10_EGEMM_KERNEL_H */
