+++ src/epblas-parallel/kind10/erot.c
@@ -1,10 +1,11 @@
+#include <stddef.h>
 /* erot — kind10 real Givens rotation. */
 typedef long double T;
 
 void erot_(const int *n_, T *x, const int *incx_, T *y, const int *incy_,
            const T *c_, const T *s_)
 {
-    const int n = *n_, incx = *incx_, incy = *incy_;
+    const ptrdiff_t n = *n_, incx = *incx_, incy = *incy_;
     const T c = *c_, s = *s_;
     if (n <= 0) return;
     /* Load both elements into locals before storing: x and y are distinct
@@ -13,15 +14,15 @@ void erot_(const int *n_, T *x, const int *incx_, T *y, const int *incy_,
      * Hoisting to locals (the epblas-openblas shape) keeps them on the x87
      * register stack. Same ops in the same order — bit-identical. */
     if (incx == 1 && incy == 1) {
-        for (int i = 0; i < n; ++i) {
+        for (ptrdiff_t i = 0; i < n; ++i) {
             T xi = x[i], yi = y[i];
             x[i] = c * xi + s * yi;
             y[i] = c * yi - s * xi;
         }
     } else {
-        int ix = (incx < 0) ? (-n + 1) * incx : 0;
-        int iy = (incy < 0) ? (-n + 1) * incy : 0;
-        for (int i = 0; i < n; ++i) {
+        ptrdiff_t ix = (incx < 0) ? (-n + 1) * incx : 0;
+        ptrdiff_t iy = (incy < 0) ? (-n + 1) * incy : 0;
+        for (ptrdiff_t i = 0; i < n; ++i) {
             T xi = x[ix], yi = y[iy];
             x[ix] = c * xi + s * yi;
             y[iy] = c * yi - s * xi;
