+++ src/epblas-parallel/kind10/escal.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 /* escal — kind10 real: X := α · X. x87 fp80, no SIMD path.
  *
  * Inner loop is 5-way unrolled to match the NETLIB DSCAL reference.
@@ -10,13 +11,13 @@ typedef long double T;
 
 void escal_(const int *n_, const T *alpha_, T *x, const int *incx_)
 {
-    const int n = *n_, incx = *incx_;
+    const ptrdiff_t n = *n_, incx = *incx_;
     const T alpha = *alpha_;
     if (n <= 0 || alpha == 1.0L) return;
     if (incx == 1) {
-        const int m = n % 5;
-        for (int i = 0; i < m; ++i) x[i] *= alpha;
-        for (int i = m; i < n; i += 5) {
+        const ptrdiff_t m = n % 5;
+        for (ptrdiff_t i = 0; i < m; ++i) x[i] *= alpha;
+        for (ptrdiff_t i = m; i < n; i += 5) {
             x[i    ] *= alpha;
             x[i + 1] *= alpha;
             x[i + 2] *= alpha;
@@ -24,7 +25,7 @@ void escal_(const int *n_, const T *alpha_, T *x, const int *incx_)
             x[i + 4] *= alpha;
         }
     } else {
-        int ix = (incx < 0) ? (-n + 1) * incx : 0;
-        for (int i = 0; i < n; ++i) { x[ix] *= alpha; ix += incx; }
+        ptrdiff_t ix = (incx < 0) ? (-n + 1) * incx : 0;
+        for (ptrdiff_t i = 0; i < n; ++i) { x[ix] *= alpha; ix += incx; }
     }
 }
